move get_status_filters to dashboard.py
